requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.*
azure-core==1.30.1
azure-cosmos==4.7.0
azure-storage-blob==12.20.0
python-dotenv==1.0.0
tenacity==8.2.3
pymongo==4.10.1
openai==0.28
flask==3.1.0
pytest